amp: update buildroot instructions for v2022.09
The method used for building an AMP payload for the Buildroot SDK has
changed for the v2022.09 release, as a Buildroot "recipe" is required to
install the remoteproc "firmware" to the filesystem. The top level
Makefile no longer handles building the examples, and a
"recipe"/sub-make is used to build the example. The choice between a
bare metal payload & a freeRTOS one has thus moved from a command line
argument to the buildroot_initramfs_defconfig.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>