Doc reorg: clean up folder structure and naming
Updated the documentation folder structure to be consistant throughout
documents with images being stored in a common image folder in the form
of:

[source section]/images/[document name]/[image]

Updated the top level folder names and organisation to be consistant
with the PolarFire SoC discussion forum.

Updated the naming convention for files so they are all lower case and
replaced "_" with "-".