apps & demos: amp: update rpmsg tty and char driver documentation
Update the rpmsg client driver documentation to align with 6.1 kernel
updates. Some of these updates include:

In 6.1 kernel, the rpmsg_char driver was split in two.
Therefore, the RPMSG_CTRL configuration must be set for backwards
compatibility.

There is an upstream rpmsg tty driver
included in the drivers/tty directory. Drop the out-of-tree
CONFIG_RPMSG_MIV_TTY in favour of the new CONFIG_RPMSG_TTY.

Previously, the /dev/ttyRPMSGx character device was given an id "x"
based on the rpmsg channel. Now with the new rpmsg tty driver, the
id is allocated using the  idr_alloc() function. This means
that the character device has changed from /dev/ttyRPMSG4 to
/dev/ttyRPMSG0.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>