apps & demos: amp: update documentation for IHC IP v2
Starting from release v2025.07 and onwards, the Icicle Kit and
Discovery Kit reference designs use the Mi-V IHC IP v2 available in the
Libero catalog.

Update the Asymmetric multiprocessing (AMP) documentation to reflect
this changes.

The Mi-V IHC IP version 2 is not backwards compatible with the
Mi-V IHC subsystem used in the Icicle Kit reference design 2025.03
or earlier. For this reason the AMP support in the v2025.07
release will not be backwards compatible with previous releases.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>