knowledge-base: remove outdated sections from mpfs-memory-configuration
These sections were only actually accurate for releases from 2022.03 to
2022.09 inclusive, and have been replaced since by overlaid memory
schemes. The early portion of the document is at least accurate in
describing the problem and the HSS portion of the Software Configuration
section is still useful. The sections relating to U-Boot, Linux and the
build systems are ultimately misleading at this point. The Buildroot SDK
has been replaced by buildroot-external and the Meta Polarfire SoC BSP
has been replaced by Meta Microchip so the instructions are invalid, and
the kernel sections invalidated by a change in how the PCI configuration
is communicated to the kernel and in what config options must be
enabled, with the "microchip,axi-m-atr0" property no longer supported.

Remove the outdated sections to avoid confusing any customers, with the
document to be updated later.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>